Delete noisy "try handler %s" log from ShellTransit
These always come up in the same order in reality, and we
only care about which handler actually played the animation.
What is more interesting is which handler we might have
skipped over, so we can log that instead of all the handlers
we offered the transition to.

Change-Id: I37c8caad27d6a2adb7a23b9cb7464705aa3417fd
Flag: EXEMPT logging change with no functional effect

Former-commit-id: cd1202d3cea10e67d16449d2201c1364be6333b4