Wait for stable window geometry in AnrTest
When AnrTest was switched to uinput, we lost the behaviour of waiting
for animations to complete prior to injection.

Likely, this caused the test to become flaky, because the activity
sometimes is not launched when we try to send input.

In this CL, we are adding a speculative fix for this by waiting for the
window geometry to stabilize before the injection starts.

Bug: 339924248
Test: atest AnrTest
Change-Id: I60ce69ce4ce8afe8b8fa3c9f48ff82c1a1ec9936