Reset mCurrentLaunchCanTurnScreenOn state after transition finishes
When the freeform Activity started, the wm shell also started the
Desktop Windowing mode and launched a DesktopWallpaperActivity on
top in the same transition, which unexpectedly resets the state.
Therefore, the freeform activity was unable to turn the screen on,
even the freeform task was also reordered to top in the same
transition.

Bug: 343450217
Test: ActivityVisibilityTests#testTurnScreenOnWithAttr_Freeform
Change-Id: I64d602770cbb5c8c82d113594f36b7894945bfc7