[Flexiglass] QQS/QS Media in landscape
Use the same treatment for both QQS/QS. Leverage the existent linear
layouts to put a view that takes half the width and that's where media
should go. We create a custom MeasurePolicy to make sure that:
* QS is measured with full width
* Media is measured with half width and located in the "end" half.

We need to do this for both QQS and QS so the animation between the
states works properly.

Some extra work is needed to make sure that the transition between QQS
and QS looks perfect, but that will be undertaken as part of the
flexiglass media transitions work.

Test: manual
Test: atest QSPanelTest
Fixes: 335826262
Flag: com.android.systemui.scene_container
Change-Id: I89d2c0f6db928c2b5e43d094c1f9feb1b3997063

Former-commit-id: c18e755393f6bbf91950092a6c903d161c65a951