Only update home visibility on exit pip to fullscreen
In ag/27443168 we update home visibility when expanding pip to
fix a bug with overview. However this introduced a new bug where
taskbar would appear when home was not visible.

This fixes the second issue by only updating home visibility when
we actually expand pip.

Flag: N/A
Bug: 342529469
Test: Manually verify taskbar doesn't appear when dismissing pip,
and that we have not regressed b/340443311 when expanding pip

Change-Id: I341118278faebf446800f89572176a3e52b5bb95
(cherry picked from commit 6c1016166e47168daf91819821be175f5831a38d)