Unused code: KeyguardController.mShowingDream
Since DreamActivity is a SHOW_WHEN_LOCKED activity, this is already
covered by checking for an occluding activity of any kind.

The DisplayPolicy definition of mShowingDream checks for an application
window belonging to ACTIVITY_TYPE_DREAM - ie. does not check for AOD
states or windowless dream - so there is no difference between these
definitions except that the KeyguardController will calculate everything
it needs at the exact time it's needed.

Flag: EXEMPT bugfix
Bug: 159821769
Bug: 337917376
Test: atest KeyguardTests
Test: atest DreamManagerServiceTests
Change-Id: I0250025c5e0bc1c7af8c0dc978f441bfbce50326