Correctly clear last placement values on nested elements
This CL ensures that we correctly clear the last placement values for
all elements that are not placed.

Before this CL, only the first element to not be placed would correctly
clear its last offset/alpha/scale given that measure() is not called on
children of elements that are not placed.

This CL fixes that by making ElementNode a traversable node that we can
recursively traverse to correctly clear the last placement values.

Bug: 290930950
Test: ElementTest
Flag: com.android.systemui.scene_container
Change-Id: I02361fcde32196ddaece01437ef727951e56d8a4