Cleanup alpha animation type after use
When entering PiP from Overview, or from legacy non-auto-enter cases, an
alpha animation is used by PipTransition. When that occurs, we tend to
clear the alpha animation type in
PipTaskOrganizer#onEndOfSwipePipToHomeTransition, which is actually not
triggered in those cases.

In turn, when a bounds animation is expected for auto entering PiP,
PipTransition#augmentRequest would think it's still in alpha animation
mode and forms the WindowContainerTransaction for bounds change,
therefore, flickers at the of auto-enter-pip animation.

Fixing this by following the same contract as in PipAnimationController,
the alpha animation type is treated as a one-shot, we reset it back to
bounds whenever it's processed.

Video: http://recall/-/aaaaaabFQoRHlzixHdtY/f3UO3cLPf5E9aNouQibRio
Bug: 341944518
Test: Enter PiP from Overview, then enter PiP again from landscape to
      portrait mode. See also the video
Change-Id: I3445b7770ef3a07e4acf6a5464b08a6e27829ca5

Former-commit-id: 4e9fc478fb393622ac06d69d20865595bf4c8d39