Fix default clock text moving right in larger display size
The reason is that when height of default clock is constrained, the font
size using half of the height is not small enough and cause TextView to
scroll when textAlignment is default. Added more detailed explanation in the
bug link.

Bug: 342505732
Test: manually test for 1) both normal and largerst display size, observe
that default clock aligned center when there's no notification, and move
left when there's notifications; 2) set system language to Arabic,
observe the symmetric behavior.
Flag: com.android.systemui.migrate_clocks_to_blueprint

Change-Id: I21e3fd04b8d656408fea822b3d176bdca6628ba6