Add signal for IME request tracking finishing
In [1] InputMethodStatsTest setup was modified to call
finishTrackingPendingImeVisibilityRequests with shell permission
identity, to ensure the required test permissions are available.
However, that is an async call, and if it takes too long the permission
could be cleared before the call is executed.

This adds a completion signal when calling the method, to ensure it
finishes before returning, maintaining the permission available.

  [1]: Ibfdc3ecea9d36ea429f98e89d0e2ee0d44857ca8

Test: atest InputMethodStatsTest
Bug: 346475189
Change-Id: I6a0d158a5e73c8a9cefcde8764c1ea7eca4fb4c8