[Sat] Change the satellite provisioned check to eager; query when start
SatelliteManager does not fire the current state for registered
callbacks at the time of registration. This change adds an
initialization step to `registerForProvisionStateChanged` such that we
always query the current status of the provisioned state and send it
through the Flow before continuing with the callback registration.

Also, we change the sharing strategy to `Eagerly` since we don't want to
generate unnecessary IPC if the device goes in and out of OOS.

Bug: 347083255
Test: DeviceBasedSatelliteRepositoryImplTest
Flag: NONE bugfix
Change-Id: Ic7025223ae803abd1b772e72fab3ccd11b175929
Merged-In: Ic7025223ae803abd1b772e72fab3ccd11b175929

Former-commit-id: 3aaaa6923faa6876e78063b6c922b4a063819f56