Improve error reporting for open-database failures
If sqlite is unable to open a database file, the root-cause is often a
Posix errno.  This CL adds the errno error message to the exception
that is thrown.  This CL also ensures that the updated error message
makes it through the exception rewrite in SQLiteConnection.open().

Manually tested with a customized build that fails with ENOSPC and
ENOENT.

Test: atest
 * FrameworksCoreTests:android.database
 * CtsDatabaseTestCases
 * SQLiteDatabasePerfTest

Bug: 338297108
Flag: EXEMPT bugfix
Change-Id: I1fdad6c95f61a01345109d70318e4531245d1f0e

Former-commit-id: 1611e110d343655f3ead1c4b8606c90b59243f1f