Implement double tap in PiP2
Implement the double tap to expand/minimize
CUJ in PiP2 through config-at-end transitions,
which should split up the full latency of communicating
with Core into two parts start and end of the animation.

Also refactor parts of the code and fix in issue with a stale state
in PipTransitionState which was causing flickers when interacting with
other CUJs.

Bug: 348723930
Flag: EXEMPT bugfix
Test: enter PiP and double tap to expand/minimize

Change-Id: Ib58e980576494f7c9fb22d98f10cb3f1925e2fa4