Fix Handle Menu window width and height
The height value in dimen was larger than the actual max height of the
menu, which meant the HandleMenu window had additional empty space after
the pills. This caused input in there (but from a user's POV, input
outside the window) to be treated as non-dimissing, which made
dismissing from that area feel unresponsive.

Additionally, the spacing (top margin) between each pill was not being
subtracted when pills were invisible.

Finally adds right and bottom padding to the view hierarchy and removes
clipping of the pills to account for the elevation that was getting
clipped by the view bounds. This requires adding that 1dp to the menu
width and height calculations before creating the
AdditionalViewContainer.

Bug: 339509495
Test: tap right below the menu's view bounds, menu is dismissed
Test: atest HandleMenuTest
Flag: EXEMPT bugfix
Change-Id: I6fefd013c7463d8690ecf9436e8c7bfe7b61744f

Former-commit-id: d2c58cc502ba6f904252772ed73c6552e384e80b