[SB][Flexi] Show lockscreen status bar even when no auth required.
KeyguardStatusBarViewModel (the model that decides whether to show the
status bar on lockscreen when scenes are enabled) was only showing the
status bar if `statusBarState == KEYGUARD`. However,
StatusBarStateController sets the statusBarState to SHADE if there's no
authentication required, even if the user is currently seeing the
lockscreen.

This CL updates KeyguardStatusBarViewModel to listen to the current
scene instead of statusBarState, since the current scene better reflects
what the user is currently seeing.

Also updates KeyguardStatusBarViewControllerTest to use more Kosmos.

Fixes: 350478931
Flag: com.android.systemui.scene_container

Test: Set device lock to just "swipe", lock device -> verify the
keyguard status bar shows when on lockscreen, but not when on AOD,
shade, or bouncer

Test: - Have face auth bypass enabled.
      - Go to the LS.
      - Receive a HUN.
      - Verify keyguard status bar hides, and status bar area just shows
	the name of the app sending a HUN
      (This verifies Ie168fe0171496d20c62057df46ce62669263e90a didn't
      regress)
Change-Id: Ia65928d16c598d9a10c22e5816abe07af6aceb37