Remove getCurrentInputMethodSubtypeForNonCurrentUsers()
This is a follow up CL to my previous CL [1], which enabled background
users to call

  InputMethodManager#getCurrentInputMethodSubtype()

by introducing

  #getCurrentInputMethodSubtypeForNonCurrentUsers().

The reason why we could not simply reuse the same implementation for

  InputMethodManagerService#getCurrentInputMethodSubtypeLocked()

was due to the fact that IMMS#mCurrentSubtype was not per-user.

Now that #mCurrentSubtype is fully per-user [2], we should be able to
unify the code.

 [1]: Icb09f9cb1a4147884faa9952b2e03ec4afa9f0b1
      5cb4f8651249d63bdb8e8930d0587c2b33af0d42
 [2]: Ibcc39212ead12c2897f8f52cb9f15b1d39959f71
      2f1bb82f41e2b8df33fd2886531f050269257092 [formerly 769e7472aad16a8a04bbbaa73a83adbf03c534ed]

Bug: 325515685
Bug: 350386877
Test: presubmit
Flag: EXEMPT refactor
Change-Id: Ia52b556ff3df3d2b50b9d62198455e5ce6ebe659

Former-commit-id: 0c32789bddfb3c58be22363521ca072a2a84d3aa