Windowing: Reparent minimizing-change to keep task behind others
Before this CL, minimizing Tasks would show in front of other Tasks
during the minimize-animation since the animation-leash is shown in
front of other Tasks. With this CL we reparent the minimizing Task to
the root Task to keep it behind other Tasks.

Bug: 335427677
Flag: com.android.window.flags.enable_desktop_windowing_task_limit
Test: minimize a Task -> shows behind other Tasks

Change-Id: I53e4d9ef76f7e291a0c4dacba64ceb71c3e98dd9

Former-commit-id: 848dac4f630ef30ec382585bb84352d7c128700d