Fix NPE in LegacyProtoLogImpl
To avoid an allocation when there are no args, the ProtoLog code processor passes null instead of the args object array.

This CL also adds a @Nullable annotation to the args parameter in the log() method, to make it clear that the parameter can be null.

Flag: EXEMPT bug fix
Bug: 351324791
Bug: 351458758
Change-Id: I2f132a4eaa42c6a22231fbb9d33204052843047e

Former-commit-id: 0d98f3df6eb6878713eabfc2577bc4d3b3cbe851