Make usages of communal hub trunk flag also read config flag
The communal hub trunk stable flag is used to control the rollout of
the feature and is enabled across products. The config flag is used
to restrict the hub to tangorpro only. This CL adds a helper method
that ANDs the two flags together and can be used to disable things that
should not run at all if the hub isn't available. This prevents
unnecessary hub logic from running outside of tangor devices.

Bug: 351058701
Fix: 351058701
Test: atest ShadeTouchHandlerTest CommunalDreamStartableTest CommunalSettingsRepositoryImplTest LegacyActivityStarterInternalImplTest FromDreamingTransitionInteractorTest FromLockscreenTransitionInteractorTest KeyguardTransitionScenariosTest CommunalOngoingContentStartableTest
Flag: com.android.systemui.communal_hub
Change-Id: Ic203bde961fb97ecfc8df01e54906a8d0cf32d79

Former-commit-id: a4f4ada8f7d70a357d214968dcb3884c0ee35350