Fix high volume of binder calls in ContentProvider.
ContentProvider.getUserIdFromAuthority can return USER_NULL when it can't parse auth as int, which can cause high volume of binder calls. This CL fixes this issue by avoiding calling into system server for invalid user ids.

This CL also fixes a bug in UserManager.getUserProperties where it can throw an exception when the user id is not supported. This CL fixes this issue by throwing an IllegalArgumentException instead.

This CL is a part of the effort to reduce the number of binder calls in the system server.

Bug: 350416200
Change-Id: Ied72302db0c141fe9ff398009331ce44a0fad2e7
Test: atest com.android.server.pm.UserManagerTest
Flag: android.multiuser.cache_user_serial_number

Former-commit-id: 65423958dec4e2fcca51826fe85e54316063050b