LockSettingsShellCommand: improve an error message
Show an appropriate error message when trying to set a lock credential
when the user already has one and the old credential wasn't provided.

Previously the empty credential got passed down to
LockPatternUtils#checkCredential(), resulting in an error of
"Credential can't be null or empty" which was confusing in this context.

Bug: 351910156
Test: locksettings set-pin 1234; locksettings set-pin 5678
Change-Id: Iaede42c5ea18fab1fc23d5c0dfd82da91c8019de