Fix the regression of Avalanche Notification when QS panel expanded
Issue: the HUN flickers when more than one HUNs are posted while the QS
panel is fully expanded (full screen). This issue was occasional before
the Avalanche throttle notification feature. The feature has a
regression which makes this issue 100% reproducible. This change fixes
the regression to make it back to the old state. For some reason, the
old HUN would disappear when a new HUN comes at the full screen QS
state, the AvalancheController prevents the showing runnable of the next
HUN from running, thus led to a wrong state. When the Shade is expanded,
including the full screen QS:

VisualStabilityCoordinator.updateAllowedStates
- mVisualStabilityProvider.setReorderingAllowed(false)

HeadsUpmanager.OnReorderingBannedListener
- Disables AvalancheController when shade is open

When the AvalancheController is disabled, we allow the runnables to run
when updateEntry, just like before the avalanche feature.

Note: the AvalancheController.delete() is not related with this issue,
I changed this to make the logic more consistent.

Bug: 340881302
Flag: com.android.systemui.notification_avalanche_throttle_hun
Test: atest SystemUITests
Change-Id: I14ca656fc76f9ae6556e9220de9e3e5c6f65f4d5