Use CommunalTransitionAnimatorController for UMO
When tapping on the UMO on the communal hub, we now wrap the animation
controller in CommunalTransitionAnimatorController in order to properly
set the communal state and hide the hub at the end of the animation.

Otherwise there is a flash at the end of the animation.

Bug: 346865769
Test: verified with UMO on hub, both over lockscreen and requiring
unlock
Flag: com.android.systemui.communal_hub

Change-Id: Ie564699846346f2e3b969afe50e81a15957f3ec4

Former-commit-id: bc6143d1332d617ff661855e5e76bfc297e593e8