Expose `Task#mLastNonFullscreenBounds` to Shell via TaskInfo
For cascading in desktop windowing, the initial bounds of the activity
are needed to be able to adjust their position but maintain the correct
size. Since the initial bounds are no longer calculated in Shell, we
we need a way of accessing the bounds calculated by Core.

The bounds calculated by core are always stored in
`Task#mLastNonFullscreenBound`. So we will add
`Task#mLastNonFullscreenBounds` to the `TaskInfo` so the correct bounds
can be accessed by Shell if and when needed.

Flag: EXEMPT adding variable
Test: Presubmit
Fixes: 355401868
Change-Id: Ib8bf1b40dcfecb9b0a92a95eda7593a7d6709f81