Introduce Content, ContentKey and ContentScope (1/2)
This CL introduces the concept of "content" in STL, which is going to be
the base class for scenes and overlays. Most of the code on this CL is
mostly a move of the implementation of what scenes are today. This will
allow to reuse most of the codes for overlays, without having to make
overlays a subclass of scenes.

Most of the existing APIs should still work. Most notably, SceneScope is
now a simple typealias to ContentScope. I decided not to deprecate it
(at least not yet) given how widespread it is. Making it a typealias of
ContentScope instead of a subclass will allow existing SceneScope.foo()
composables to be directly reusable in overlays.

See go/sysui-stl-overlay for details about overlays.

Bug: 353679003
Test: atest PlatformComposeSceneTransitionLayoutTests
Flag: com.android.systemui.scene_container

Change-Id: I9f84a5bd518fe0409171477dd484ee60def22d70

Former-commit-id: 259a4f3421a82af670596737444ce0d3efe2af01