Fix UserHandle
LowLuxModifier, DMD and DPC register observer while the system is starting (current user = system), but on some devices, later the current user switches. When Smooth Display for example is enabled with the UI, it's enabled by the new user and not the system user. We should therefore use USER_ALL when registering observers.

When reading the settings, we should use USER_CURRENT instead of getting the user id from the context associated with the system.

We should also re-read the settings if the user switches.

Additionally, update the stale evenDimmer comment in DDC.

Bug: 353645971
Test: AutomaticBrightnessControllerTest, BrightnessClamperControllerTest, BrightnessLowLuxModifierTest, BrightnessMappingStrategyTest, DisplayManagerServiceTest, DisplayModeDirectorTest, DisplayPowerControllerTest
Flag: EXEMPT bugfix
Change-Id: I204478b6f356aea72f50bf9130f92e53b2758202