Reconnect "full width" shade logic with scenes enabled
QS, ScrimController, and NSSL all have a boolean field for whether the
shade is using the entire width of the device. At this time, this
directly corresponds to whether the device is not fullscreen. When the
scene container is enabled, there is no NPVC to set these values, so
this change collects from a flow and set them.

Test: added new tests and ran affected tests
Fixes: 347739712
Fixes: 349333683
Flag: com.android.systemui.scene_container
Change-Id: Ia5066859a7fc67fd6a3477420e71951f4258ffe5

Former-commit-id: 672db05430d112505dce997655585abe3d70620c