Creates API for QS that avoids getHeader
As we are migrating to use Compose inside the fragment, we need an API
that will not require QS to have an actual header view. This new API
extracts all of the usages of QS.getHeader and replaces them (flagged)
with APIs that directly query the necessary values.

Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Bug: 353254345
Test: atest QSImplTest QuickSettingsControllerImplTest
Test: atest NotificationStackScrollLayoutTest
Change-Id: Ie0425848a57943e8b6799d574427a1c7cfe73b9e

Former-commit-id: 1d13c765fcf62f18b7753624b2cff680a91fe7de