Remove UserLifecycleListener deps from SecureSettingsWrapper (2nd try)
This is the 2nd attempt to submit my original CL [1], which got
reverted [2] to verify whether it was causing jank benchmark
regression discussed in Bug 355334507 or not. As reverting my original
CL did not affect the benchmark, I am now submitting this again.

As part of removing the dependency on UserLifecycleListener from
random utility classes under com.android.server.inputmethod.*, this CL
updates SecureSettingsWrapper so that it does not need to register
UserLifecycleListener by itself.

This is a mechanical refactoring CL. The observable semantics should
remain unchanged.

 [1]: If903fe15e7e71eed2fccb45c082d8efb8176f26c
      fcf67e16ce89b39d9881102377459d6699920122 [formerly 1a7311fe9a04a840f730144d72f12d74f2005511]
 [2]: Ib4ce3cb283f5165cff0e66c9a6ddc29910abd911
      671c7888ac50fbdad6424423b311d259a44d78e9 [formerly 44a859a61eebc24f995c73eb0af96abfe82cf176]

Fix: 352354308
Test: presubmit
Flag: EXEMPT refactor
Change-Id: I7f1502d19df95e6ba5228dd8b501e7e98d210cfd

Former-commit-id: 332eb8c1c7be3a8b31cc3425d98870cad17f7257