Move LaunchAdjacentController logic from Desktop* into FreeformListener.
Currently the controller is set to disable LaunchAdjacent if and only if
Android Desktop Mode is enabled, and there are freeform tasks visible.
Instead of disabling it only when ADM is on, change the callback so that
when there's any >0 freeform tasks appearing in FreeformTaskListener,
diasble LaunchAdjacent.

Bug: 357649930
Test: Manual; In a non-ADM device, open Chrome w/ Freeform window,
then New Window - see that a new freeform window appears
Flag: EXEMPT Bug fix

Change-Id: I615f93a580e8a0b27cbc30b5947815bfcd69c7be

Former-commit-id: e63f933b816092d93879b2222cf721538c2c3bf7